Last fixes strategy, our moves are random
